fix sniffer and composer.json

+++ src/Command/StorageInfoCommand.php
@@ -126,7 +126,7 @@ class StorageInfoCommand extends Command
         $units = ['B', 'KB', 'MB', 'GB', 'TB'];
         $bytes = max($bytes, 0);
         $pow = floor(($bytes ? log($bytes) : 0) / log(1024));
-        $pow = min($pow, count($units) - 1);
+        $pow = (int)min($pow, count($units) - 1);
         $bytes /= (1 << (10 * $pow));
 
         return round($bytes, 2) . ' ' . $units[$pow];
